run_perf_test: fix skewed values caused by stderr
Previously, any stderr would be written to result_comp and result_decomp.  Since
throughput is calculated with `awk ... result_comp`, stderr messages can
vastly skew the reported compthroughput value. This problem is also
theoretically possible for decompthroughput, but it was not observed.

We fix this by only writing stdout to result_comp and result_decomp, NOT stderr.
Instead, stderr is written to result_comp_stderr/result_decomp_stderr, which
are deleted when this script is invoked.